requirements.txt
pycairo
pyyaml
pyserial
